Implement blueprint for users
