+++ website/database/credentials.json
@@ -5,20 +5,20 @@
             "password": "admin"
         },
         {
-            "username": "Jo\u00e3o Carlos",
-            "password": "Jo\u00e3o Carlos"
-        },
-        {
             "username": "Que site lindo",
             "password": "mt pika"
         },
         {
-            "username": "Serio",
-            "password": "fazer trabalho com vcs \u00e9 mt foda"
-        },
-        {
             "username": "Que lindo cara",
             "password": "amo vcs"
+        },
+        {
+            "username": "serio",
+            "password": "serio"
+        },
+        {
+            "username": "Jo\u00e3o Carlos",
+            "password": "Jo\u00e3o Carlos"
         }
     ]
 }
