+++ website/flask_data.py
@@ -1,0 +1,6 @@
+# Settings
+debug_mode = True # Enable this to be able to view pages all pages without logging in
+
+# Data
+current_user = ""
+admin_mode = False
