+++ website/templates/home_page.html
@@ -29,7 +29,7 @@
             {% if admin_mode %}
               <li>
                 <!-- <i class="ph-fill ph-users"></i> -->
-                <a href="{{ url_for('list_user') }}">Usuários</a>
+                <a href="{{ url_for('user_blueprint.list_user') }}">Usuários</a>
               </li>
               <li>
                 <!-- <i class="ph-fill ph-train"></i> -->
