+++ website/templates/list_user.html
@@ -15,7 +15,7 @@
       <article> <!-- Replace with a base later! -->
         <p>{{admin_name}}</p>
         <div class="tools">
-          <form id="form" method="POST" action="{{ url_for('edit_user') }}">
+          <form id="form" method="POST" action="{{ url_for('user_blueprint.edit_user') }}">
           <button type="submit" class="edit-icon" id=""><!--  colocar id-->
             <i class="ph ph-pencil-line"></i>
           </button>
@@ -27,13 +27,13 @@
       <article>
         <p>{{user}}</p>
         <div class="tools">
-          <form id="form" method="POST" action="{{ url_for('edit_user') }}">
+          <form id="form" method="POST" action="{{ url_for('user_blueprint.edit_user') }}">
             <button type="submit" class="edit-icon" id=""><!--  colocar id-->
               <i class="ph ph-pencil-line"></i>
             </button>
             <input type="hidden" name="user_id" id="user_id" value="{{id}}">
           </form>
-          <form id="form" method="POST" action="{{ url_for('try_remove_user') }}">
+          <form id="form" method="POST" action="{{ url_for('user_blueprint.try_remove_user') }}">
           <button type="submit" class="trash-icon" id=""><!--  colocar id-->
             <i class="ph ph-trash"></i>
           </button>
@@ -46,7 +46,7 @@
     </div>
   </div>
   <div class="credencial">
-    <a href="{{url_for('register_user')}}">
+    <a href="{{url_for('user_blueprint.register_user')}}">
       <button type="submit" id=""> <!--  colocar id-->
         Cadastrar usuários
       </button>
